minor: OpenBlt remove type conversions

+++ java_console/io/src/main/java/com/rusefi/libopenblt/XcpLoader.java
@@ -9,19 +9,19 @@ import java.util.Arrays;
 
 public class XcpLoader {
     // XCP command codes as defined by the protocol currently supported by this module
-    private static final int XCPLOADER_CMD_CONNECT = 0xFF;    // XCP connect command code
-    private static final int XCPLOADER_CMD_GET_STATUS = 0xFD;    // XCP get status command code
-    private static final int XCPLOADER_CMD_GET_SEED = 0xF8;    // XCP get seed command code
-    private static final int XCPLOADER_CMD_UNLOCK = 0xF7;    // XCP unlock command code
-    private static final int XCPLOADER_CMD_SET_MTA = 0xF6;    // XCP set mta command code
-    private static final int XCPLOADER_CMD_UPLOAD = 0xF5;    // XCP upload command code
-    private static final int XCPLOADER_CMD_PROGRAM_START = 0xD2;    // XCP program start command code
-    private static final int XCPLOADER_CMD_PROGRAM_CLEAR = 0xD1;    // XCP program clear command code
-    private static final int XCPLOADER_CMD_PROGRAM = 0xD0;    // XCP program command code
-    private static final int XCPLOADER_CMD_PROGRAM_RESET = 0xCF;    // XCP program reset command code
-    private static final int XCPLOADER_CMD_PROGRAM_MAX = 0xC9;    // XCP program max command code
+    private static final byte XCPLOADER_CMD_CONNECT = (byte)0xFF;    // XCP connect command code
+    private static final byte XCPLOADER_CMD_GET_STATUS = (byte)0xFD;    // XCP get status command code
+    private static final byte XCPLOADER_CMD_GET_SEED = (byte)0xF8;    // XCP get seed command code
+    private static final byte XCPLOADER_CMD_UNLOCK = (byte)0xF7;    // XCP unlock command code
+    private static final byte XCPLOADER_CMD_SET_MTA = (byte)0xF6;    // XCP set mta command code
+    private static final byte XCPLOADER_CMD_UPLOAD = (byte)0xF5;    // XCP upload command code
+    private static final byte XCPLOADER_CMD_PROGRAM_START = (byte)0xD2;    // XCP program start command code
+    private static final byte XCPLOADER_CMD_PROGRAM_CLEAR = (byte)0xD1;    // XCP program clear command code
+    private static final byte XCPLOADER_CMD_PROGRAM = (byte)0xD0;    // XCP program command code
+    private static final byte XCPLOADER_CMD_PROGRAM_RESET = (byte)0xCF;    // XCP program reset command code
+    private static final byte XCPLOADER_CMD_PROGRAM_MAX = (byte)0xC9;    // XCP program max command code
 
-    private static final int XCPLOADER_CMD_PID_RES = 0xFF;   // positive response
+    private static final byte XCPLOADER_CMD_PID_RES = (byte)0xFF;   // positive response
 
     private static final int XCPLOADER_PACKET_SIZE_MAX = 255;
 
@@ -59,7 +59,7 @@ public class XcpLoader {
     private void connect() throws IOException {
         byte[] request =
             ByteBuffer.allocate(2)
-                .put((byte) XCPLOADER_CMD_CONNECT)
+                .put(XCPLOADER_CMD_CONNECT)
                 .put((byte) 0)  // connection mode?
                 .array();
 
@@ -69,7 +69,7 @@ public class XcpLoader {
             throw new IOException("connect() failed due to incorrect length received: " + response.length);
         }
 
-        if (response[0] != (byte) XCPLOADER_CMD_PID_RES) {
+        if (response[0] != XCPLOADER_CMD_PID_RES) {
             throw new IOException("connect() failed due to bad response ID received: " + response[0]);
         }
 
@@ -103,7 +103,7 @@ public class XcpLoader {
     private int programStart() throws IOException {
         byte[] request =
             ByteBuffer.allocate(1)
-                .put((byte) XCPLOADER_CMD_PROGRAM_START)
+                .put(XCPLOADER_CMD_PROGRAM_START)
                 .array();
 
         byte[] response = mTransport.sendPacket(request, mSettings.timeoutT6, 7);
@@ -112,7 +112,7 @@ public class XcpLoader {
             throw new IOException("sendCmdProgramStart() failed due to incorrect length received: " + response.length);
         }
 
-        if (response[0] != (byte) XCPLOADER_CMD_PID_RES) {
+        if (response[0] != XCPLOADER_CMD_PID_RES) {
             throw new IOException("sendCmdProgramStart() failed due to bad response ID received: " + response[0]);
         }
 
@@ -137,7 +137,7 @@ public class XcpLoader {
         byte[] request =
             ByteBuffer.allocate(8)
                 .order(mTargetInfo.byteOrder)
-                .put((byte)XCPLOADER_CMD_PROGRAM_CLEAR)
+                .put(XCPLOADER_CMD_PROGRAM_CLEAR)
                 .put((byte)0)
                 .put((byte)0)
                 .put((byte)0)
@@ -146,7 +146,7 @@ public class XcpLoader {
 
         byte[] response = mTransport.sendPacket(request, mSettings.timeoutT4, 1);
 
-        if (response.length != 1 || response[0] != (byte)XCPLOADER_CMD_PID_RES) {
+        if (response.length != 1 || response[0] != XCPLOADER_CMD_PID_RES) {
             throw new IOException("programClear failed for address=" + address + " length=" + len);
         }
     }
@@ -171,14 +171,14 @@ public class XcpLoader {
     private void sendCmdProgram(byte[] data) throws IOException {
         byte[] request =
             ByteBuffer.allocate(2 + data.length)
-                .put((byte)XCPLOADER_CMD_PROGRAM)
+                .put(XCPLOADER_CMD_PROGRAM)
                 .put((byte)data.length)
                 .put(data)
                 .array();
 
         byte[] response = mTransport.sendPacket(request, mSettings.timeoutT5, 1);
 
-        if (response.length != 1 || response[0] != (byte)XCPLOADER_CMD_PID_RES) {
+        if (response.length != 1 || response[0] != XCPLOADER_CMD_PID_RES) {
             throw new IOException("sendCmdProgram failed");
         }
     }
@@ -186,7 +186,7 @@ public class XcpLoader {
     private void sendCmdProgramReset() throws IOException {
         byte[] request =
             ByteBuffer.allocate(1)
-                .put((byte) XCPLOADER_CMD_PROGRAM_RESET)
+                .put(XCPLOADER_CMD_PROGRAM_RESET)
                 .array();
 
         byte[] response;
@@ -202,7 +202,7 @@ public class XcpLoader {
             throw new IOException("sendCmdProgramReset() failed due to incorrect length received: " + response.length);
         }
 
-        if (response[0] != (byte) XCPLOADER_CMD_PID_RES) {
+        if (response[0] != XCPLOADER_CMD_PID_RES) {
             throw new IOException("sendCmdProgramReset() failed due to bad response ID received: " + response[0]);
         }
     }
@@ -211,7 +211,7 @@ public class XcpLoader {
         byte[] request =
             ByteBuffer.allocate(8)
                 .order(mTargetInfo.byteOrder)
-                .put((byte)XCPLOADER_CMD_SET_MTA)
+                .put(XCPLOADER_CMD_SET_MTA)
                 .put((byte)0)
                 .put((byte)0)
                 .put((byte)0)
@@ -220,7 +220,7 @@ public class XcpLoader {
 
         byte[] response = mTransport.sendPacket(request, mSettings.timeoutT1, 1);
 
-        if (response.length != 1 || response[0] != (byte)XCPLOADER_CMD_PID_RES) {
+        if (response.length != 1 || response[0] != XCPLOADER_CMD_PID_RES) {
             throw new IOException("setMta failed for address " + address);
         }
     }
